Catching up on commits.  Various changes, largly logging.
Was not in the habit of frequent commits and am playing catchup.

I don't know all the changes, but looking through the diff seems
that many changes were made, mainly bug fixes. I believe the
largest change from the last commit is adding logging
functionality to a text file in AppData.

I promise to be better with my version control for this project
moving forward.