Refactor on close event to properly clean up timers if close canceled.
Previous timer cleanup was done on WorkbookBeforeClose, but this
cleaned up the timer even if the close was canceled.  Therefore,
the WorkbookBeforeClose handler was replaced with clean up on
WorkbookDeactivate.  Now whenever any workbook is deactivated,
the list of open timers is checked against the list of open workbooks.
If any timer doesn't have a corresponding open workbook, it is
removed.

This means that the exposed CloseTimer method now actually cleans up
ALL timers that no longer have a corresponding open workbook.